Rename to sfarkxtc
sfark is the name of the original linux sfark console decompression utility.

In actual fact this code is based on sfarkxtm: the code for sfarkxtc has
been lost to history, but it too was based on sfarkxtm, the mac version of
this utility, which was in turn based on the windows tool sfarkxt, all based
on the sfArkLib library.

This version is intended to replace both sfarkxtm and sfarkxtc and be
cross-platform.